telegram-bot/requirements.txt
python-telegram-bot==21.9
httpx==0.28.1
python-dotenv==1.0.1
structlog==24.4.0
